:bug: Refactore path for ModalWindow
